[algebra.Function] fix dtype if `derive` returns the zero polynomial

+++ pySecDec/algebra.py
@@ -268,7 +268,7 @@ class Function(_Expression):
         if summands:
             derivative = self.derivatives[index] = Sum(*summands, copy=False)
         else: # return zero packed into a `Polynomial`
-            derivative = self.derivatives[index] = Polynomial(np.zeros([1,self.number_of_variables]), np.array([0]), self.symbols, copy=False)
+            derivative = self.derivatives[index] = Polynomial(np.zeros([1,self.number_of_variables], dtype=int), np.array([0]), self.symbols, copy=False)
 
         return derivative
 
@@ -490,7 +490,7 @@ class Polynomial(_Expression):
                 return summand1 + summand2 # implicit simplify
         else:
             if summand1 is None:
-                return Polynomial(np.zeros([1,self.number_of_variables],dtype=int), np.array([0]), self.polysymbols, copy=False)
+                return Polynomial(np.zeros([1,self.number_of_variables], dtype=int), np.array([0]), self.polysymbols, copy=False)
             else:
                 return summand1
